admin-backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-JWT-Extended==4.5.3
PyMySQL==1.1.0
python-dotenv==1.0.0
bcrypt==4.0.1
cryptography==41.0.4
flask-mail
python-dotenv
flask-cors
Flask-Limiter
